+++ servers/quarkus-tests/src/main/java/org/projectnessie/quarkus/tests/profiles/QuarkusTestProfilePersistMySQL.java
@@ -15,7 +15,7 @@
  */
 package org.projectnessie.quarkus.tests.profiles;
 
-import static org.projectnessie.quarkus.config.VersionStoreConfig.VersionStoreType.JDBC;
+import static org.projectnessie.quarkus.config.VersionStoreConfig.VersionStoreType.JDBC2;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -28,7 +28,7 @@ public class QuarkusTestProfilePersistMySQL extends BaseConfigProfile {
   public Map<String, String> getConfigOverrides() {
     return ImmutableMap.<String, String>builder()
         .putAll(super.getConfigOverrides())
-        .put("nessie.version.store.type", JDBC.name())
+        .put("nessie.version.store.type", JDBC2.name())
         .put("nessie.version.store.persist.jdbc.datasource", "mysql")
         .put("quarkus.datasource.mysql.jdbc.extended-leak-report", "true")
         .build();