+++ versioned/storage/jdbc/src/main/java/org/projectnessie/versioned/storage/jdbc/JdbcBackendBaseConfig.java
@@ -20,24 +20,4 @@ import java.util.Optional;
 public interface JdbcBackendBaseConfig {
 
   Optional<String> datasourceName();
-
-  /**
-   * The JDBC catalog name.
-   *
-   * @deprecated This setting has never worked as expected and is now ineffective. The catalog must
-   *     be specified directly in the JDBC URL using the option {@code
-   *     quarkus.datasource.*.jdbc.url}.
-   */
-  @Deprecated(forRemoval = true)
-  Optional<String> catalog();
-
-  /**
-   * The JDBC schema name.
-   *
-   * @deprecated This setting has never worked as expected and is now ineffective. The schema must
-   *     be specified directly in the JDBC URL using the option {@code
-   *     quarkus.datasource.*.jdbc.url}.
-   */
-  @Deprecated(forRemoval = true)
-  Optional<String> schema();
 }